# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraSol API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "description": "Get all Permissions",
                "operationId": "getPermissionsList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Store new Permission",
                "description": "Returns permission data",
                "operationId": "storePermisions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePermissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permission information",
                "description": "Returns permission data",
                "operationId": "getPermissionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update existing permission",
                "description": "Returns updated permission data",
                "operationId": "updatePermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePermissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Permission"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Delete existing permission",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePermission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePermissionRequest": {
                "title": "Store Permission Form request",
                "required": [
                    "id",
                    "county_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "county_name": {
                        "description": "County Name",
                        "type": "string",
                        "example": "Nairobi"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StorePermissionRequest"
                }
            },
...
            "UpdatePermissionRequest": {
                "title": "Update Permission Form request",
                "required": [
                    "id",
                    "name",
                    "guard_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name",
                        "type": "string"
                    },
                    "guard_name": {
                        "description": "Guard name",
                        "type": "string",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdatePermissionRequest"
                }
            },
...
            "Permission": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "guard_name": {
                        "description": "Guard name",
                        "type": "string",
                        "readOnly": "true"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Permission"
                }
            },
...
        }
    }
}